Create isUserInfo Loading context
