+++ src/features/common/users/hooks/index.ts
@@ -8,22 +8,29 @@ export const UserContext = createContext<{
   setUserInfo: (info: User | null) => void;
   loginModalType: LoginModalType;
   setLoginModalType: (type: LoginModalType) => void;
+  isUserInfoLoading: boolean;
+  setIsUserInfoLoading: (status: boolean) => void
 }>({
   userInfo: null,
   setUserInfo: () => {},
   loginModalType: '',
-  setLoginModalType: () => {}
+  setLoginModalType: () => {},
+  isUserInfoLoading: false,
+  setIsUserInfoLoading: () => {}
 });
 
 const useUserInfo = () => {
   const [userInfo, setUserInfo] = useState<User | null>(null)
   const [loginModalType, setLoginModalType] = useState<LoginModalType>('')
+  const [isUserInfoLoading, setIsUserInfoLoading] = useState(false)
 
   return {
     userInfo,
     setUserInfo,
     loginModalType,
-    setLoginModalType
+    setLoginModalType,
+    isUserInfoLoading,
+    setIsUserInfoLoading
   }
 }
