+++ src/pages/_app.tsx
@@ -66,12 +66,7 @@ export default function App({ Component, pageProps }: AppProps) {
 
   const [isPageLoading, setIsPageLoading] = useState(false);
 
-  const {
-    userInfo,
-    setUserInfo,
-    loginModalType,
-    setLoginModalType
-  } = useUserInfo()
+  const userInfoContextValue = useUserInfo()
 
   useEffect(() => {
     const routeEventStart = () => {
@@ -96,14 +91,7 @@ export default function App({ Component, pageProps }: AppProps) {
     <>
       <GlobalStyle />
       <Provider store={store}>
-        <UserContext.Provider
-          value={{
-            userInfo,
-            setUserInfo,
-            loginModalType,
-            setLoginModalType
-          }}
-        >
+        <UserContext.Provider value={userInfoContextValue}>
           <ThemeProvider theme={theme}>
             <MUIThemeProvider theme={muiDarkTheme}>
               {isPageLoading && <PageLoading />}
